feat : Ajout des domaines swagger.io et editor.swagger.io à la liste des domaines acceptés par CORS + Suppression fichiers inutiles
