+++ api-core.net/Controllers/ParticipationController.cs
@@ -7,16 +7,30 @@ using System.Threading.Tasks;
 
 namespace api_core.net.Controllers
 {
+    /**
+     * Controller de la classe Participation
+     * La route de base du controller est :
+     * quiz/{idQuiz}/participation
+     * idQuiz, représente l'id du quiz auquel la participation se rapporte
+     **/
     [Route("quiz/{idQuiz}/participation")]
     public class ParticipationController : Controller
     {
+        // Le DAO nécessaire au traitement des données
         ParticipationDao participationDao;
 
         public ParticipationController(ParticipationDao participationDao)
         {
             this.participationDao = participationDao;
         }
-        //Tâche pour envoyer idQuiz et participation
+
+        /**
+         * Route POST
+         * Créé un objet Participation et l'URL de cette dernière
+         * HTTP Code : 201
+         * 
+         * @param p un objet Participation
+         **/
         [HttpPost]
         public async Task<IActionResult> Post(string idQuiz, [FromBody]Participation p)
         {
@@ -25,7 +39,14 @@ namespace api_core.net.Controllers
 
             return new CreatedResult($"/quiz/{idQuiz}/participation/{p.Id}", p);
         }
-        //Mise a jour des données d'une participation
+
+        /**
+         * Route PUT
+         * Edite un objet Participation et renvoie sur la page de cette dernière
+         * HTTP Code : 200
+         * 
+         * @param p un objet Participation
+         **/
         [HttpPut("{idParticipation}")]
         public async Task<IActionResult> Put(string idQuiz, string idParticipation, [FromBody]Participation p)
         {
@@ -42,7 +63,15 @@ namespace api_core.net.Controllers
             await participationDao.Update(idPart, p);
             return new OkResult();
         }
-        //Permet de récupérer l'Identifiant
+
+        /**
+         * Route GET
+         * Renvoie un objet Participation 
+         * HTTP Code : 200
+         * 
+         * @param idParticipation id de la participation
+         * @return un objet Participation
+         **/
         [HttpGet("{idParticipation}")]
         public IActionResult GetId(string idParticipation)
         {
@@ -54,8 +83,15 @@ namespace api_core.net.Controllers
             return new ObjectResult(participation);
         }
 
+        /**
+         * Route PATCH
+         * PATCH un objet participation
+         * HTTP Code : 200
+         * 
+         * @param idParticipation id de la participation
+         * @param patch l'objet participation sous forme de JSONPatchDocument
+         **/
         [HttpPatch("{idParticipation}")]
-        //Mise à jour de l'identifiant participation
         public async Task<IActionResult> Patch(string idParticipation, [FromBody]JsonPatchDocument<Participation> patch)
         {
             var participation = participationDao.GetParticipation(new ObjectId(idParticipation));
