+++ api-core.net/Controllers/QuizController.cs
@@ -8,16 +8,30 @@ using System.Threading.Tasks;
 
 namespace api_core.net.Controllers
 {
+    /**
+     * Controller de la classe Quiz
+     * La route de base du controller est :
+     * quiz
+     **/
     [Route("quiz")]
     public class QuizController : Controller
     {
+        // Le DAO nécessaire au traitement des données
         QuizDao quizDao;
 
         public QuizController(QuizDao quizDao)
         {
             this.quizDao = quizDao;
         }
-        //Obtenir l'identifiant du Quiz
+
+        /**
+         * Route GET
+         * Renvoie un objet Quiz 
+         * HTTP Code : 200
+         * 
+         * @param idQuiz id du quiz
+         * @return un objet Quiz
+         **/
         [HttpGet("{idQuiz}")]
         public IActionResult GetId(string idQuiz)
         {
@@ -28,7 +42,14 @@ namespace api_core.net.Controllers
             }
             return new ObjectResult(quiz);
         }
-        //Obtenir tous les quizz
+
+        /**
+         * Route GET
+         * Renvoie tous les quiz 
+         * HTTP Code : 200
+         * 
+         * @return list d'objet Quiz
+         **/
         [HttpGet]
         public IActionResult GetAll()
         {
@@ -40,6 +61,13 @@ namespace api_core.net.Controllers
             return new ObjectResult(listQuiz);
         }
 
+        /**
+         * Route POST
+         * Créé un objet Quiz et l'URL de cette dernière
+         * HTTP Code : 201
+         * 
+         * @param quiz un objet Quiz
+         **/
         [HttpPost]
         public async Task<IActionResult> Post([FromBody]Quiz quiz)
         {
@@ -47,7 +75,15 @@ namespace api_core.net.Controllers
 
             return new CreatedResult($"/quiz/{quiz.Id}", quiz);
         }
-        //Mettre à jour l'identifiant de quiz
+
+        /**
+         * Route PATCH
+         * PATCH un objet quiz
+         * HTTP Code : 200
+         * 
+         * @param idQuiz id de la quiz
+         * @param patch l'objet quiz sous forme de JSONPatchDocument
+         **/
         [HttpPatch("{idQuiz}")]
         public async Task<IActionResult> Patch(string idQuiz, [FromBody]JsonPatchDocument<Quiz> patch)
         {
