--- api-core.net/Dockerfile
@@ -1,6 +1,0 @@
-FROM api-core.net-base
-ADD /src /app
-WORKDIR /app/api-core.net
-RUN ["dotnet", "restore"]
-EXPOSE 80
-ENTRYPOINT ["dotnet", "run"]
