+++ api-core.net/Models/Question.cs
@@ -5,6 +5,12 @@ using System.Collections.Generic;
 
 namespace api_core.net.Models
 {
+    /**
+     * Modélisation d'une question :
+     * - Un titre
+     * - Une image, si la question est en image
+     * - Une liste de réponses
+     **/
     public class Question
     {
         [BsonElement("title")]
