+++ api-core.net/Startup.cs
@@ -1,11 +1,11 @@
 ﻿using api_core.net.Daos;
-using api_core.net.Models;
 using api_core.net.Utils;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.AspNetCore.Cors;
 
 namespace api_core.net
 {
@@ -36,10 +36,23 @@ namespace api_core.net
         // This method gets called by the runtime. Use this method to add services to the container
         public void ConfigureServices(IServiceCollection services)
         {
+            // Enable CORS 
+            services.AddCors(options =>
+            {
+                // Add the right domains
+                options.AddPolicy("Swagger",
+                    builder => builder.WithOrigins("http://swagger.io")
+                    .AllowAnyHeader());
+                options.AddPolicy("SwaggerEditor",
+                    builder => builder.WithOrigins("http://editor.swagger.io")
+                    .AllowAnyHeader());
+            });
+
             // Add framework services.
             services.AddApplicationInsightsTelemetry(Configuration);
             services.AddTransient<QuizDao>();
             services.AddTransient<ParticipationDao>();
+
             // Add JsonConverter for ObjectId
             services.AddMvc().AddJsonOptions(opt => { opt.SerializerSettings.Converters.Add(new ObjectIdConverter()); });
         }
@@ -49,6 +62,11 @@ namespace api_core.net
         {
             loggerFactory.AddConsole(Configuration.GetSection("Logging"));
             loggerFactory.AddDebug();
+
+            // Add swagger domain to CORS   
+            app.UseCors("Swagger");
+            app.UseCors("SwaggerEditor");
+
 
             app.UseApplicationInsightsRequestTelemetry();
