--- api-core.net/api-core.net.sln
@@ -1,32 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{58058F47-496A-4DF9-A2C6-FD2D9850E2D3}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{6166D5B4-27CD-4F29-80EA-C0002D88F40D}"
-	ProjectSection(SolutionItems) = preProject
-		global.json = global.json
-	EndProjectSection
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "api-core.net", "src\api-core.net\api-core.net.xproj", "{6781EBA6-09EC-41E2-9061-516FBF8FCE2E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6781EBA6-09EC-41E2-9061-516FBF8FCE2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6781EBA6-09EC-41E2-9061-516FBF8FCE2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6781EBA6-09EC-41E2-9061-516FBF8FCE2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6781EBA6-09EC-41E2-9061-516FBF8FCE2E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{6781EBA6-09EC-41E2-9061-516FBF8FCE2E} = {58058F47-496A-4DF9-A2C6-FD2D9850E2D3}
-	EndGlobalSection
-EndGlobal
