--- api-core.net/global.json
@@ -1,6 +1,0 @@
-﻿{
-  "projects": [ "src", "test" ],
-  "sdk": {
-    "version": "1.0.0-preview2-003131"
-  }
-}
